Cache symtab across images in dyld shared cache (#183)
* Use new inclusive/exclusive load command filters

A CacheImage can now load the segments it needs on-demand by checking
which ones have been loaded already. There is no longer a need for a
partial Mach-O anymore; it can be upgraded to a full one at any time.

* Cache symtab across images in dyld shared cache

In a shared cache, all images have a strtab that is shared and very
large. We should keep a single copy of this around and reuse it instead
of copying it for each image.